+++ juwon/core/core/src/main/java/hello/core/member/Member.java
