+++ juwon/core/core/src/main/java/hello/core/member/MemberServiceImpl.java
@@ -2,10 +2,7 @@ package hello.core.member;
 
 public class MemberServiceImpl implements MemberService {
 
-    private final MemberRepository memberRepository;
-    public MemberServiceImpl(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
+    private final MemberRepository memberRepository=new MemoryMemberRepository();
 
     @Override
     public void join(Member member){
@@ -14,7 +11,6 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public Member findMember(Long memberId) {
-
         return memberRepository.findById(memberId);
     }
 }
