+++ juwon/core/core/src/main/java/hello/core/order/OrderServiceImpl.java
@@ -2,23 +2,14 @@ package hello.core.order;
 
 import hello.core.discount.DiscountPolicy;
 import hello.core.discount.FixDiscountPolicy;
-import hello.core.discount.RateDiscountPolicy;
 import hello.core.member.Member;
 import hello.core.member.MemberRepository;
 import hello.core.member.MemoryMemberRepository;
 
 public class OrderServiceImpl implements OrderService{
 
-    //DIP 지키고 있다.
-    private final MemberRepository memberRepository;
- //   private final DiscountPolicy discountPolicy=new FixDiscountPolicy();
-    //private final DiscountPolicy discountPolicy = new RateDiscountPolicy();
-    private DiscountPolicy discountPolicy;
-
-    public OrderServiceImpl(MemberRepository memberRepository, DiscountPolicy discountPolicy){
-        this.memberRepository=memberRepository;
-        this.discountPolicy=discountPolicy;
-    }
+    private final MemberRepository memberRepository=new MemoryMemberRepository();
+    private final DiscountPolicy discountPolicy=new FixDiscountPolicy();
 
     @Override
     public Order createOrder(Long memberId, String itemName, int itemPrice) {
