+++ juwon/core/core/src/main/resources/application.properties
