+++ juwon/core/core/src/test/java/hello/core/member/MemberServiceTest.java
@@ -1,18 +1,11 @@
 package hello.core.member;
 
-import hello.core.AppConfig;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class MemberServiceTest {
 
-    MemberService memberService;
-    @BeforeEach
-    public void beforeEach() {
-        AppConfig appConfig = new AppConfig();
-        memberService = appConfig.memberService();
-    }
+    MemberService memberService=new MemberServiceImpl();
 
     @Test
     void join(){
