+++ juwon/core/core/src/test/java/hello/core/order/OrderServiceTest.java
@@ -1,24 +1,15 @@
 package hello.core.order;
 
-import hello.core.AppConfig;
 import hello.core.member.*;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static hello.core.member.Grade.VIP;
 
 public class OrderServiceTest {
 
-    MemberService memberService;
-    OrderService orderService;
-
-    @BeforeEach
-    public void beforeEach() {
-        AppConfig appConfig = new AppConfig();
-        memberService = appConfig.memberService();
-        orderService = appConfig.orderService();
-    }
+    MemberService memberService=new MemberServiceImpl();
+    OrderService orderService=new OrderServiceImpl();
 
     @Test
     void createOrder(){
