--- juwon/core/src/main/java/hello/core/MemberApp.java
@@ -1,30 +1,0 @@
-package hello.core;
-
-import hello.core.member.Grade;
-import hello.core.member.Member;
-import hello.core.member.MemberService;
-import org.springframework.context.ApplicationContext;
-import
-        org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-public class MemberApp {
-
-    public static void main(String[] args) {
-
-        // AppConfig appConfig = new AppConfig();
-        // MemberService memberService = appConfig.memberService();
-        ApplicationContext applicationContext = new
-                AnnotationConfigApplicationContext(AppConfig.class);
-
-        MemberService memberService =
-                applicationContext.getBean("memberService", MemberService.class);
-
-        Member member = new Member(1L, "memberA", Grade.VIP);
-
-        memberService.join(member);
-
-        Member findMember = memberService.findMember(1L);
-        System.out.println("new member = " + member.getName());
-        System.out.println("find Member = " + findMember.getName());
-    }
-}
