--- juwon/core/src/main/java/hello/core/OrderApp.java
@@ -1,33 +1,0 @@
-package hello.core;
-
-import hello.core.member.Grade;
-import hello.core.member.Member;
-import hello.core.member.MemberService;
-import hello.core.order.Order;
-import hello.core.order.OrderService;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-public class OrderApp {
-
-    public static void main(String[] args) {
-
-        // AppConfig appConfig = new AppConfig();
-        // MemberService memberService = appConfig.memberService();
-        // OrderService orderService = appConfig.orderService();
-        ApplicationContext applicationContext = new
-                AnnotationConfigApplicationContext(AppConfig.class);
-
-        MemberService memberService =
-                applicationContext.getBean("memberService", MemberService.class);
-
-        OrderService orderService = applicationContext.getBean("orderService",
-                OrderService.class);
-
-        long memberId = 1L;
-        Member member = new Member(memberId, "memberA", Grade.VIP);
-        memberService.join(member);
-        Order order = orderService.createOrder(memberId, "itemA", 10000);
-        System.out.println("order = " + order);
-    }
-}
