--- juwon/core/src/main/java/hello/core/discount/RateDiscountPolicy.java
@@ -1,16 +1,0 @@
-package hello.core.discount;
-
-import hello.core.member.Grade;
-import hello.core.member.Member;
-
-public class RateDiscountPolicy implements DiscountPolicy {
-    private int discountPercent = 10; //10% 할인
-    @Override
-    public int discount(Member member, int price) {
-        if (member.getGrade() == Grade.VIP) {
-            return price * discountPercent / 100;
-        } else {
-            return 0;
-        }
-    }
-}
